requirements.txt
# UI Framework
PySide6

# Database ORM
SQLAlchemy

# Data Engine
polars

# API & Web Requests
requests
python-dotenv

# Testing
pytest

#Widgets
matplotlib